# runtime-acm/src/main/resources/openapi/openapi.yaml
...
openapi: 3.0.1
info:
  title: Api Documentation
  description: Api Documentation
  termsOfService: urn:tos
  contact: {}
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0
  version: "1.0"
servers:
  - url: https://{server}
    variables:
      server:
        default: onap/acm/v3
        description: This value is assigned by the service provider
...
paths:
...
  /compositions/{compositionId}/instances:
    get:
      tags:
      - Automation Composition Instance
      summary: Query details of the requested automation composition instances
      description: Query details of the requested automation composition instances for the given automation composition definition ID, returning
        details of all its automation composition instances
      operationId: queryCompositionInstances
      parameters:
      - name : compositionId
        in: path
        description: The UUID of the automation composition definition for which to return instances
        required: true
        schema:
          type: string
          format: uuid
      - name: name
        in: query
        description: Automation composition instance name. Regular expressions are supported for filtering. If
          this parameter is not specified, all automation composition instances for the specified definition are returned.
        schema:
          type: string
      - name: version
        in: query
        description: Automation composition instance version. Regular expressions are supported for filtering. If this
          parameter is not specified, all automation composition instances for the specified definition that match the "name" filter are are returned.
        schema:
          type: string
      - name: X-onap-RequestId
        in: header
        description: RequestID for http transaction
        schema:
          type: string
          format: uuid
      responses:
        200:
          description: Serialised instance of
            [AutomationCompositions](https://github.com/onap/policy-clamp/blob/master/models/src/main/java/org/onap/policy/clamp/models/acm/concepts/AutomationCompositions.java)
            containing a list of automation composition instances found
          headers:
            X-LatestVersion:
              schema:
                type: string
            X-PatchVersion:
              schema:
                type: string
            X-MinorVersion:
              schema:
                type: string
            X-onap-RequestId:
              schema:
                type: string
                format: uuid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AutomationCompositions'
              example:
                externalValue: 'https://raw.githubusercontent.com/onap/policy-clamp/master/runtime-acm/src/main/resources/openapi/examples/getCompositionInstancesResponse.json'
            application/yaml:
              schema:
                $ref: '#/components/schemas/AutomationCompositions'
              example:
                externalValue: 'https://raw.githubusercontent.com/onap/policy-clamp/master/runtime-acm/src/main/resources/openapi/examples/getCompositionInstancesResponse.yaml'
        401:
          description: Authentication Error, returns an instance of
            [SimpleResponse](https://github.com/onap/policy-clamp/blob/master/models/src/main/java/org/onap/policy/clamp/models/acm/messages/rest/SimpleResponse.java)
          headers:
            X-LatestVersion:
              schema:
                type: string
            X-PatchVersion:
              schema:
                type: string
            X-MinorVersion:
              schema:
                type: string
            X-onap-RequestId:
              schema:
                type: string
                format: uuid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SimpleResponse'
        404:
          description: The specified automation composition definition was not found, returns an instance of
            [SimpleResponse](https://github.com/onap/policy-clamp/blob/master/models/src/main/java/org/onap/policy/clamp/models/acm/messages/rest/SimpleResponse.java)
          headers:
            X-LatestVersion:
              schema:
                type: string
            X-PatchVersion:
              schema:
                type: string
            X-MinorVersion:
              schema:
                type: string
            X-onap-RequestId:
              schema:
                type: string
                format: uuid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SimpleResponse'
      security:
      - basicAuth: []
      x-interface info:
        api-version: 1.0.0
        last-mod-release: London
    post:
      tags:
      - Automation Composition Instance
      summary: Create automation composition instance
      description: Creates automation composition instance that use the sepcified automation composition definition. The ID of the created
        automation composition instance is returned.
      operationId: createCompositionInstance
      parameters:
      - name : compositionId
        in: path
        description: The UUID of the automation composition definition on which to create instance
        required: true
        schema:
          type: string
          format: uuid
      - name: X-onap-RequestId
        in: header
        description: RequestID for http transaction
        schema:
          type: string
          format: uuid
      requestBody:
        description: Serialised instance of
          [AutomationComposition](https://github.com/onap/policy-clamp/blob/master/models/src/main/java/org/onap/policy/clamp/models/acm/concepts/AutomationComposition.java)
          containing a automation composition instance to create
        content:
          application/json:
            schema:
                $ref: '#/components/schemas/AutomationComposition'
            example:
              externalValue: 'https://raw.githubusercontent.com/onap/policy-clamp/master/runtime-acm/src/main/resources/openapi/examples/postCompositionInstance.json'
          application/yaml:
            schema:
                $ref: '#/components/schemas/AutomationComposition'
            example:
              externalValue: 'https://raw.githubusercontent.com/onap/policy-clamp/master/runtime-acm/src/main/resources/openapi/examples/postCompositionInstance.yaml'
        required: true
      responses:
        201:
          description: Serialised instance of
            [InstantiationResponse](https://github.com/onap/policy-clamp/blob/master/models/src/main/java/org/onap/policy/clamp/models/acm/messages/rest/instantiation/InstantiationResponse.java)
            containing the UUID of the created automation composition instance
          headers:
            X-LatestVersion:
              schema:
                type: string
            X-PatchVersion:
              schema:
                type: string
            X-MinorVersion:
              schema:
                type: string
            X-onap-RequestId:
              schema:
                type: string
                format: uuid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InstantiationResponse'
              example:
                externalValue: 'https://raw.githubusercontent.com/onap/policy-clamp/master/runtime-acm/src/main/resources/openapi/examples/postCompositionInstancesResponse.json'
            application/yaml:
              schema:
                $ref: '#/components/schemas/InstantiationResponse'
              example:
                externalValue: 'https://raw.githubusercontent.com/onap/policy-clamp/master/runtime-acm/src/main/resources/openapi/examples/postCompositionInstancesResponse.yaml'
        401:
          description: Authentication Error, returns an instance of
            [SimpleResponse](https://github.com/onap/policy-clamp/blob/master/models/src/main/java/org/onap/policy/clamp/models/acm/messages/rest/SimpleResponse.java)
          headers:
            X-LatestVersion:
              schema:
                type: string
            X-PatchVersion:
              schema:
                type: string
            X-MinorVersion:
              schema:
                type: string
            X-onap-RequestId:
              schema:
                type: string
                format: uuid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SimpleResponse'
        404:
          description: The specified automation composition definition was not found, returns an instance of
            [SimpleResponse](https://github.com/onap/policy-clamp/blob/master/models/src/main/java/org/onap/policy/clamp/models/acm/messages/rest/SimpleResponse.java)
          headers:
            X-LatestVersion:
              schema:
                type: string
            X-PatchVersion:
              schema:
                type: string
            X-MinorVersion:
              schema:
                type: string
            X-onap-RequestId:
              schema:
                type: string
                format: uuid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SimpleResponse'
        400:
          description: Bad Request, returns an instance of
            [SimpleResponse](https://github.com/onap/policy-clamp/blob/master/models/src/main/java/org/onap/policy/clamp/models/acm/messages/rest/SimpleResponse.java)
          headers:
            X-LatestVersion:
              schema:
                type: string
            X-PatchVersion:
              schema:
                type: string
            X-MinorVersion:
              schema:
                type: string
            X-onap-RequestId:
              schema:
                type: string
                format: uuid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SimpleResponse'
      security:
      - basicAuth: []
      x-interface info:
        api-version: 1.0.0
        last-mod-release: London
      x-codegen-request-body-name: body
  /compositions/{compositionId}/instances/{instanceId}:
    get:
      tags:
      - Automation Composition Instance
      summary: Get automation composition instance details.
      description: Get details of the requested automation composition instance.
      operationId: getCompositionInstance
      parameters:
      - name : compositionId
        in: path
        description: The UUID of the automation composition definition on which to get an instance
        required: true
        schema:
          type: string
          format: uuid
      - name : instanceId
        in: path
        description: The UUID of the automation composition instance to get
        required: true
        schema:
          type: string
          format: uuid
      - name: X-onap-RequestId
        in: header
        description: RequestID for http transaction
        schema:
          type: string
          format: uuid
      responses:
        200:
          description: Serialised instance of
            [AutomationComposition](https://github.com/onap/policy-clamp/blob/master/models/src/main/java/org/onap/policy/clamp/models/acm/concepts/AutomationComposition.java)
            containing the automation composition instance
          headers:
            X-LatestVersion:
              schema:
                type: string
            X-PatchVersion:
              schema:
                type: string
            X-MinorVersion:
              schema:
                type: string
            X-onap-RequestId:
              schema:
                type: string
                format: uuid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AutomationComposition'
              example:
                externalValue: 'https://raw.githubusercontent.com/onap/policy-clamp/master/runtime-acm/src/main/resources/openapi/examples/getCompositionInstanceResponse.json'
            application/yaml:
              schema:
                $ref: '#/components/schemas/AutomationComposition'
              example:
                externalValue: 'https://raw.githubusercontent.com/onap/policy-clamp/master/runtime-acm/src/main/resources/openapi/examples/getCompositionInstanceResponse.yaml'
        401:
          description: Authentication Error, returns an instance of
            [SimpleResponse](https://github.com/onap/policy-clamp/blob/master/models/src/main/java/org/onap/policy/clamp/models/acm/messages/rest/SimpleResponse.java)
          headers:
            X-LatestVersion:
              schema:
                type: string
            X-PatchVersion:
              schema:
                type: string
            X-MinorVersion:
              schema:
                type: string
            X-onap-RequestId:
              schema:
                type: string
                format: uuid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SimpleResponse'
        404:
          description: The automation composition instance was not found, returns an instance of
            [SimpleResponse](https://github.com/onap/policy-clamp/blob/master/models/src/main/java/org/onap/policy/clamp/models/acm/messages/rest/SimpleResponse.java)
          headers:
            X-LatestVersion:
              schema:
                type: string
            X-PatchVersion:
              schema:
                type: string
            X-MinorVersion:
              schema:
                type: string
            X-onap-RequestId:
              schema:
                type: string
                format: uuid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SimpleResponse'
      security:
      - basicAuth: []
      x-interface info:
        api-version: 1.0.0
        last-mod-release: London
    put:
      tags:
      - Automation Composition Instance
      summary: Manage deployment and locking of an automation composition instance
      description: This request manages deployment and locking of an automation composition instance. This endpoint can
        order deployment and undeployment of an AC Instance to participants and order unlocking and locking of AC instances
        on participants
      operationId: ompositionInstanceState
      parameters:
      - name : compositionId
        in: path
        description: The UUID of the automation composition definition on which to update an instance
        required: true
        schema:
          type: string
          format: uuid
      - name : instanceId
        in: path
        description: The UUID of the automation composition instance to update
        required: true
        schema:
          type: string
          format: uuid
      - name: X-onap-RequestId
        in: header
        description: RequestID for http transaction
        schema:
          type: string
          format: uuid
      requestBody:
        description: Serialised instance of
            [AcInstanceStateUpdate](https://github.com/onap/policy-clamp/blob/master/models/src/main/java/org/onap/policy/clamp/models/acm/messages/rest/instantiation/AcInstanceStateUpdate.java)
            which specifies the requested state change on the automation concept instance
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AcInstanceStateUpdate'
            example:
              externalValue: 'https://raw.githubusercontent.com/onap/policy-clamp/master/runtime-acm/src/main/resources/openapi/examples/putAcInstanceStateUpdate.json'
          application/yaml:
            schema:
              $ref: '#/components/schemas/AcInstanceStateUpdate'
            example:
              externalValue: 'https://raw.githubusercontent.com/onap/policy-clamp/master/runtime-acm/src/main/resources/openapi/examples/putAcInstanceStateUpdate.yaml'
        required: true
      responses:
        202:
          description: Accepted, the request has been accepted and forwarded to participants
          headers:
            X-LatestVersion:
              schema:
                type: string
            X-PatchVersion:
              schema:
                type: string
            X-MinorVersion:
              schema:
                type: string
            X-onap-RequestId:
              schema:
                type: string
                format: uuid
        400:
          description: Bad Request, returns an instance of
            [SimpleResponse](https://github.com/onap/policy-clamp/blob/master/models/src/main/java/org/onap/policy/clamp/models/acm/messages/rest/SimpleResponse.java)
          headers:
            X-LatestVersion:
              schema:
                type: string
            X-PatchVersion:
              schema:
                type: string
            X-MinorVersion:
              schema:
                type: string
            X-onap-RequestId:
              schema:
                type: string
                format: uuid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SimpleResponse'
        401:
          description: Authentication Error, returns an instance of
            [SimpleResponse](https://github.com/onap/policy-clamp/blob/master/models/src/main/java/org/onap/policy/clamp/models/acm/messages/rest/SimpleResponse.java)
          headers:
            X-LatestVersion:
              schema:
                type: string
            X-PatchVersion:
              schema:
                type: string
            X-MinorVersion:
              schema:
                type: string
            X-onap-RequestId:
              schema:
                type: string
                format: uuid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SimpleResponse'
        404:
          description: The specified automation composition instance was not found, returns an instance of
            [SimpleResponse](https://github.com/onap/policy-clamp/blob/master/models/src/main/java/org/onap/policy/clamp/models/acm/messages/rest/SimpleResponse.java)
          headers:
            X-LatestVersion:
              schema:
                type: string
            X-PatchVersion:
              schema:
                type: string
            X-MinorVersion:
              schema:
                type: string
            X-onap-RequestId:
              schema:
                type: string
                format: uuid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SimpleResponse'
      security:
      - basicAuth: []
      x-interface info:
        api-version: 1.0.0
        last-mod-release: London
      x-codegen-request-body-name: body
    delete:
      tags:
      - Automation Composition Instance
      summary: Delete an automation composition instance
      description: Deletes a automation composition instance, returning the UUID of the deleted automation composition instance
      operationId: deleteCompositionInstance
      parameters:
      - name : compositionId
        in: path
        description: The UUID of the automation composition definition on which to delete an instance
        required: true
        schema:
          type: string
          format: uuid
      - name : instanceId
        in: path
        description: The UUID of the automation composition instance to delete
        required: true
        schema:
          type: string
          format: uuid
      - name: X-onap-RequestId
        in: header
        description: RequestID for http transaction
        schema:
          type: string
          format: uuid
      responses:
        200:
          description: Serialised instance of
            [InstantiationResponse](https://github.com/onap/policy-clamp/blob/master/models/src/main/java/org/onap/policy/clamp/models/acm/messages/rest/instantiation/InstantiationResponse.java)
            containing the UUID of the deleted automation composition instance
          headers:
            X-LatestVersion:
              schema:
                type: string
            X-PatchVersion:
              schema:
                type: string
            X-MinorVersion:
              schema:
                type: string
            X-onap-RequestId:
              schema:
                type: string
                format: uuid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InstantiationResponse'
              example:
                externalValue: 'https://raw.githubusercontent.com/onap/policy-clamp/master/runtime-acm/src/main/resources/openapi/examples/deleteCompositionInstanceResponse.json'
            application/yaml:
              schema:
                $ref: '#/components/schemas/InstantiationResponse'
              example:
                externalValue: 'https://raw.githubusercontent.com/onap/policy-clamp/master/runtime-acm/src/main/resources/openapi/examples/deleteCompositionInstanceResponse.yaml'
        401:
          description: Authentication Error, returns an instance of
            [SimpleResponse](https://github.com/onap/policy-clamp/blob/master/models/src/main/java/org/onap/policy/clamp/models/acm/messages/rest/SimpleResponse.java)
          headers:
            X-LatestVersion:
              schema:
                type: string
            X-PatchVersion:
              schema:
                type: string
            X-MinorVersion:
              schema:
                type: string
            X-onap-RequestId:
              schema:
                type: string
                format: uuid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SimpleResponse'
        404:
          description: The specified automation composition instance was not found, returns an instance of
            [SimpleResponse](https://github.com/onap/policy-clamp/blob/master/models/src/main/java/org/onap/policy/clamp/models/acm/messages/rest/SimpleResponse.java)
          headers:
            X-LatestVersion:
              schema:
                type: string
            X-PatchVersion:
              schema:
                type: string
            X-MinorVersion:
              schema:
                type: string
            X-onap-RequestId:
              schema:
                type: string
                format: uuid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SimpleResponse'
        400:
          description: Bad Request, returns an instance of
            [SimpleResponse](https://github.com/onap/policy-clamp/blob/master/models/src/main/java/org/onap/policy/clamp/models/acm/messages/rest/SimpleResponse.java)
          headers:
            X-LatestVersion:
              schema:
                type: string
            X-PatchVersion:
              schema:
                type: string
            X-MinorVersion:
              schema:
                type: string
            X-onap-RequestId:
              schema:
                type: string
                format: uuid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SimpleResponse'
      security:
      - basicAuth: []
      x-interface info:
        api-version: 1.0.0
        last-mod-release: London
components:
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic
  schemas:
...
    AutomationComposition:
      title: AutomationComposition
      type: object
    AutomationCompositions:
      title: AutomationCompositions
      type: object
    SimpleResponse:
      title: SimpleResponse
      type: object
...
    AcInstanceStateUpdate:
      title: AcInstanceStateUpdate
      type: object
    InstantiationResponse:
      title: InstantiationResponse
      type: object